% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currstatCIR_copula_saved_nuisances.R
\name{currstatCIR_copula_saved_nuisances}
\alias{currstatCIR_copula_saved_nuisances}
\title{Estimate a survival function under current status sampling with copula}
\usage{
currstatCIR_copula_saved_nuisances(
  time,
  event,
  X,
  mu_n,
  g_n,
  Riemann_grid,
  F_sIx_n,
  eval_region,
  n_eval_pts = 101,
  theta = 0.5
)
}
\description{
Estimate a survival function under current status sampling with copula
}
