% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_w_stack.R
\name{f_w_stack}
\alias{f_w_stack}
\title{Wrapper for various f_w stacked algorithms}
\usage{
f_w_stack(
  time,
  event,
  X,
  entry,
  censored,
  bin_size = NULL,
  algorithm,
  V = NULL,
  SL.library = NULL,
  time_basis,
  tuning_params = NULL,
  direction = "forward"
)
}
\arguments{
\item{time}{Observed time}

\item{event}{Indicator of event (vs censoring)}

\item{X}{Covariate matrix}

\item{entry}{Variable indicating time of entry into the study (truncation variable) if applicable}

\item{censored}{Logical, indicates whether to run regression on censored observations (vs uncensored)}

\item{bin_size}{Size of time bin on which to discretize for estimation}

\item{algorithm}{Which binary classification algorithm to use}

\item{V}{CV fold number, required for tuned algorithms (xgboost, ranger)}

\item{SL.library}{SuperLearner library}

\item{time_basis}{How to treat time (continuous or dummy)}

\item{tuning_params}{Tuning parameters for binary classification}
}
\value{
An fitted pooled binary regression for the truncation distribution
}
\description{
Wrapper for various f_w stacked algorithms
}
