% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_y_algorithms.R
\name{f_y_ecdf}
\alias{f_y_ecdf}
\title{Empirical cdf}
\usage{
f_y_ecdf(time, event, X, censored)
}
\arguments{
\item{censored}{}
}
\value{
An object of class \code{f_y_ecdf}
}
\description{
Empirical cdf
}
