% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_y_algorithms.R
\name{f_y_listWrappers}
\alias{f_y_listWrappers}
\title{List wrappers for f_y prediction}
\usage{
\code{f_y_listWrappers()}
}
\value{
Invisible character vector with the requested functions.
}
\description{
This function lists all prediction algorithms built into the \code{conSurv} package.
}
