% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebuilt_generators.R
\name{generate_nuisance_predictions_stackG}
\alias{generate_nuisance_predictions_stackG}
\title{Estimate conditional survival function nuisance parameters using survival stacking}
\usage{
generate_nuisance_predictions_stackG(
  time,
  event,
  X,
  X_holdout,
  newtimes,
  SL.library = c("SL.mean", "SL.glm", "SL.earth", "SL.gam", "SL.ranger"),
  V = 5,
  bin_size = 0.05,
  approx_times
)
}
\arguments{
\item{time}{\code{n x 1} numeric vector of observed
follow-up times. If there is censoring, these are the minimum of the
event and censoring times.}

\item{event}{\code{n x 1} numeric vector of status indicators of
whether an event was observed.}

\item{X}{\code{n x p} data.frame of observed covariate values}

\item{X_holdout}{\code{m x p} data.frame of new observed covariate
values at which to obtain \code{m} predictions for the estimated algorithm.
Must have the same names and structure as \code{X}.}

\item{newtimes}{\code{k x 1} numeric vector of times at which to obtain \code{k}
predicted conditional survivals.}

\item{SL.library}{Super Learner library}

\item{V}{Number of cross-validation folds, to be passed to \code{SuperLearner}}

\item{bin_size}{Size of time bin on which to discretize for estimation
of cumulative probability functions. Can be a number between 0 and 1,
indicating the size of quantile grid (e.g. \code{0.1} estimates
the cumulative probability functions on a grid based on deciles of
observed \code{time}s). If \code{NULL}, creates a grid of
all observed \code{time}s. See \link{stackG} documentation.}

\item{approx_times}{Numeric vector of times at which to
approximate product integral or cumulative hazard interval. See \link{stackG} documentation.}
}
\value{
A list containing elements \code{S_hat} (conditional event survival function, corresponding to \code{X_holdout} and \code{newtimes}),
\code{S_hat_train} (conditional event survival function, corresponding to \code{X} and \code{newtimes}),
\code{G_hat} (conditional censoring survival function, corresponding to \code{X_holdout} and \code{newtimes}),
and \code{G_hat_train} (conditional censoring survival function, corresponding to \code{X} and \code{newtimes})
}
\description{
Estimate conditional survival function nuisance parameters using survival stacking
}
\seealso{
\link{stackG}
}
