% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebuilt_generators.R
\name{generate_oracle_predictions_boost}
\alias{generate_oracle_predictions_boost}
\title{Estimate oracle prediction function using DR gradient boosting}
\usage{
generate_oracle_predictions_boost(
  time,
  event,
  X,
  X_holdout,
  nuisance_preds,
  restriction_time,
  approx_times,
  V = 5,
  indx,
  tuning = FALSE,
  subsample_n = length(time),
  boosting_params = list(mstop = c(100), nu = c(0.1), sigma = c(0.01), learner =
    c("glm"))
)
}
\arguments{
\item{time}{\code{n x 1} numeric vector of observed
follow-up times. If there is censoring, these are the minimum of the
event and censoring times.}

\item{event}{\code{n x 1} numeric vector of status indicators of
whether an event was observed.}

\item{X}{\code{n x p} data.frame of observed covariate values}

\item{X_holdout}{\code{m x p} data.frame of new observed covariate
values at which to obtain \code{m} predictions for the estimated algorithm.
Must have the same names and structure as \code{X}.}

\item{nuisance_preds}{Named list of conditional survival function predictions with elements \code{"S_hat"}, \code{"S_hat_train"},
\code{"G_hat"}, and \code{"G_hat_train"}. This should match the output of \code{conditional_surv_generator}.}

\item{restriction_time}{Maximum follow-up time for calculation of the C-index.
Essentially, this time should be chosen such that the conditional survival function is identified at
this time for all covariate values \code{X} present in the data. Choosing the restriction time such that roughly 10\% of individuals remain at-risk
at that time has been shown to work reasonably well in simulations.}

\item{approx_times}{Numeric vector of length J2 giving times at which to
approximate C-index integral.}

\item{V}{Number of cross-validation folds for selection of tuning parameters}

\item{indx}{Numeric index of column(s) of \code{X} to be removed, i.e., not used in the oracle prediction function.}

\item{tuning}{Logical, whether or not to use cross-validation to select tuning parameters}

\item{subsample_n}{Number of samples to use for boosting procedure. Using a subsample of the full sample can greatly reduce runtime}

\item{boosting_params}{Named list of parameter values for the boosting procedure. Elements of this list include \code{mstop} (number of
boosting iterations), \code{nu} (learning rate), \code{sigma} (smoothness parameter for sigmoid approximation, with smaller meaning
less smoothing), and \code{learner} (base learner, can take values \code{"glm"}, \code{"gam"}, or \code{"tree"})}
}
\value{
A list containing elements \code{f0_hat} and \code{f0_hat_train}, the estimated oracle prediction functions for
\code{X_holdout} and \code{X}, respectively.
}
\description{
Estimate oracle prediction function using DR gradient boosting
}
\seealso{
\link{boost_c_index}
}
