% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prebuilt_generators.R
\name{generate_oracle_predictions_SL}
\alias{generate_oracle_predictions_SL}
\title{Estimate small oracle prediction function using Super Learner regression}
\usage{
generate_oracle_predictions_SL(
  time,
  event,
  X,
  X_holdout,
  nuisance_preds,
  outcome,
  landmark_times,
  restriction_time,
  approx_times,
  SL.library = c("SL.mean", "SL.glm", "SL.earth", "SL.gam", "SL.ranger"),
  V = 5,
  indx
)
}
\arguments{
\item{time}{\code{n x 1} numeric vector of observed
follow-up times. If there is censoring, these are the minimum of the
event and censoring times.}

\item{event}{\code{n x 1} numeric vector of status indicators of
whether an event was observed.}

\item{X}{\code{n x p} data.frame of observed covariate values}

\item{X_holdout}{\code{m x p} data.frame of new observed covariate
values at which to obtain \code{m} predictions for the estimated algorithm.
Must have the same names and structure as \code{X}.}

\item{nuisance_preds}{Named list of conditional survival function predictions with elements \code{"S_hat"}, \code{"S_hat_train"},
\code{"G_hat"}, and \code{"G_hat_train"}. This should match the output of \code{conditional_surv_generator}.}

\item{outcome}{Outcome type, either \code{"survival_probability"} or \code{"restricted_survival_time"}}

\item{landmark_times}{Numeric vector of length J1 giving
landmark times at which to estimate VIM (\code{"accuracy"}, \code{"AUC"}, \code{"Brier"}, \code{"R-squared"}).}

\item{restriction_time}{Maximum follow-up time for calculation of \code{"survival_time_MSE"}. Essentially, this time
should be chosen such that the conditional survival function is identified at this time for all covariate values \code{X} present in the data.
Choosing the restriction time such that roughly 10\% of individuals remain at-risk at that time has been shown to work reasonably well in simulations.}

\item{approx_times}{Numeric vector of length J2 giving times at which to
approximate integral appearing in the pseudo-outcomes}

\item{SL.library}{Super Learner library}

\item{V}{Number of cross-validation folds, to be passed to \code{SuperLearner}}

\item{indx}{Numeric index of column(s) of \code{X} to be removed, i.e., not used in the oracle prediction function.}
}
\value{
A list containing elements \code{f0_hat} and \code{f0_hat_train}, the estimated small oracle prediction functions for
\code{X_holdout} and \code{X}, respectively.
}
\description{
When the oracle prediction function is a conditional mean, the small oracle prediction function can be estimated by
regressing the large oracle prediction function on the small feature vector. This function performs this regression using
Super Learner.
}
