% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridSurv.R
\name{gridSurv}
\alias{gridSurv}
\title{Estimate a conditional survival function by binary regression on a grid}
\usage{
gridSurv(
  time,
  event,
  X,
  newX,
  newtimes,
  time_grid_approx,
  bin_size = NULL,
  denom_method = "stratified",
  algorithm = "xgboost",
  V = 10,
  entry = NULL,
  time_basis,
  surv_form = "PI",
  tuning_params = NULL
)
}
\arguments{
\item{time}{Observed time}

\item{event}{Indicator of event (vs censoring)}

\item{X}{Covariate matrix}

\item{newX}{Values of covariates at which to make a prediction}

\item{newtimes}{Times at which to make the survival function prediction}

\item{time_grid_approx}{Grid of time points on which to approximate PL integral}

\item{algorithm}{Which binary classification algorithm to use (xgboost, ranger, gam with smoothing splines)}

\item{V}{CV fold number, required for tuned algorithms (xgboost, ranger)}

\item{entry}{Variable indicating time of entry into the study (truncation variable)}

\item{test_time}{Observed times to evaluate against}

\item{test_event}{Event indicators to test against}

\item{test_X}{Covariates corresponding to \code{test_times} and \code{test_event}}
}
\value{
An object of class \code{gridSurv}
}
\description{
Estimate a conditional survival function by binary regression on a grid
}
