% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernSurv.R
\name{kernSurv}
\alias{kernSurv}
\title{Estimate a conditional survival function using kernel regression for cdfs}
\usage{
kernSurv(
  time,
  event,
  X,
  newX,
  newtimes,
  time_grid_approx,
  kernel_type,
  kernel_order,
  SL.library,
  denom_method = "stratified"
)
}
\arguments{
\item{time}{Observed time}

\item{event}{Indicator of event (vs censoring)}

\item{X}{Covariate matrix}

\item{newX}{Values of covariates at which to make a prediction}

\item{newtimes}{Times at which to make the survival function prediction}

\item{time_grid_approx}{Grid of time points on which to approximate PL integral}

\item{SL.library}{SuperLearner library}

\item{denom_method}{Method of computing the denominator}
}
\value{
An object of class \code{kernSurv}
}
\description{
Estimate a conditional survival function using kernel regression for cdfs
}
