% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_delta.R
\name{p_delta}
\alias{p_delta}
\title{Wrapper for various p_delta algorithms}
\usage{
p_delta(event, X, V = 10, SL.library)
}
\arguments{
\item{event}{\code{n x 1} numeric vector of status indicators of
whether an event was observed. Defaults to a vector of 1s, i.e. no censoring.}

\item{X}{\code{n x p} data.frame of observed covariate values
on which to train the estimator.}

\item{V}{Number of cross validation folds on which to train the Super Learner
classifier. Defaults to 10.}

\item{SL.library}{Library of algorithms to include in the binary classification
Super Learner. Should have the same structure as the \code{SL.library}
argument to the \code{SuperLearner} function in the \code{SuperLearner} package.}
}
\value{
An fitted binary regression for (complement of)
probability of censoring
}
\description{
Wrapper for various p_delta algorithms
}
