% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_delta_listWrappers.R
\name{p_delta_listWrappers}
\alias{p_delta_listWrappers}
\title{List wrappers for p_delta prediction}
\usage{
\code{p_delta_listWrappers()}
}
\value{
Invisible character vector with the requested functions.
}
\description{
This function lists all prediction algorithms built into the \code{conSurv} package.
}
