% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survML-package.R
\docType{package}
\name{survML-package}
\alias{survML}
\alias{survML-package}
\title{survML: Flexible Estimation of Conditional Survival Functions Using Machine Learning}
\description{
Tools for flexible estimation of conditional survival functions using off-the-shelf machine learning tools. Implements both global and local survival stacking. See Wolock CJ, Gilbert PB, Simon N, and Carone M (2022+) \href{https://arxiv.org/abs/2211.03031}{arXiv:2211.03031}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cwolock/survML}
  \item Report bugs at \url{https://github.com/cwolock/survML/issues}
}

}
\author{
\strong{Maintainer}: Charles Wolock \email{cwolock@gmail.com} (\href{https://orcid.org/0000-0003-3527-1102}{ORCID}) [copyright holder]

}
\keyword{internal}
