% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survMLs.R
\name{survMLs}
\alias{survMLs}
\title{Estimate a conditional survival function via stacking}
\usage{
survMLs(
  time,
  event,
  X,
  newX,
  newtimes,
  bin_size = NULL,
  V = 10,
  time_basis = "continuous",
  algorithm = "xgboost",
  entry = NULL,
  direction = "prospective",
  tuning_params = NULL,
  SL.library = NULL
)
}
\arguments{
\item{time}{Observed time}

\item{event}{Indicator of event (vs censoring)}

\item{X}{Covariate matrix}

\item{newX}{Values of covariates at which to make a prediction}

\item{newtimes}{Times at which to make the survival function prediction}

\item{bin_size}{Quantiles on which to grid times. If NULL, defaults to every observed time}

\item{V}{Number of cross validation folds}

\item{time_basis}{How to treat time (continuous or dummy)}

\item{algorithm}{Binary classification algorithm to use}

\item{entry}{Study entry variable, if applicable}

\item{direction}{Prospective or retrospective study}

\item{tuning_params}{Tuning parameters to be passed to algorithm for tuning}

\item{SL.library}{SuperLearner library}
}
\value{
An object of class \code{survMLs}
}
\description{
Estimate a conditional survival function via stacking
}
