% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vim.R
\name{vim}
\alias{vim}
\title{Estimate AUC VIM}
\usage{
vim(
  type,
  time,
  event,
  X,
  landmark_times = quantile(time[event == 1], probs = c(0.25, 0.5, 0.75)),
  restriction_time = max(time[event == 1]),
  approx_times = NULL,
  large_feature_vector,
  small_feature_vector,
  conditional_surv_preds = NULL,
  large_oracle_preds = NULL,
  small_oracle_preds = NULL,
  conditional_surv_generator = NULL,
  conditional_surv_generator_control = NULL,
  large_oracle_generator = NULL,
  large_oracle_generator_control = NULL,
  small_oracle_generator = NULL,
  small_oracle_generator_control = NULL,
  cf_folds = NULL,
  cf_fold_num = 5,
  sample_split = TRUE,
  ss_folds = NULL,
  robust = TRUE,
  scale_est = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{type}{Type of VIM to compute}

\item{time}{\code{n x 1} numeric vector of observed
follow-up times If there is censoring, these are the minimum of the
event and censoring times.}

\item{event}{\code{n x 1} numeric vector of status indicators of
whether an event was observed. Defaults to a vector of 1s, i.e. no censoring.}

\item{landmark_times}{Numeric vector of length J2 giving
times at which to estimate AUC}

\item{restriction_time}{Maximum follow-up time for calculation of C-index or restricted survival time}

\item{approx_times}{Numeric vector of length J1 giving times at which to
approximate integrals.}

\item{cf_folds}{Numeric vector of length n giving cross-fitting folds}

\item{cf_fold_num}{The number of cross-fitting folds, if not providing \code{cf_folds}}

\item{sample_split}{Logical indicating whether or not to sample split}

\item{ss_folds}{Numeric vector of length n giving sample-splitting folds}

\item{robust}{Logical, whether or not to use the doubly-robust debiasing approach. This option
is meant for illustration purposes only --- it should be left as \code{TRUE}.}

\item{scale_est}{Logical, whether or not to force the VIM estimate to be nonnegative}

\item{alpha}{The level at which to compute confidence intervals and hypothesis tests. Defaults to 0.05}
}
\value{
data frame giving results
}
\description{
Estimate AUC VIM
}
