% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vim_AUC.R
\name{vim_AUC}
\alias{vim_AUC}
\title{Estimate AUC VIM}
\usage{
vim_AUC(
  time,
  event,
  approx_times,
  landmark_times,
  f_hat,
  fs_hat,
  S_hat,
  G_hat,
  folds,
  sample_split,
  ss_folds,
  robust,
  scale_est = FALSE
)
}
\arguments{
\item{time}{\code{n x 1} numeric vector of observed
follow-up times If there is censoring, these are the minimum of the
event and censoring times.}

\item{event}{\code{n x 1} numeric vector of status indicators of
whether an event was observed. Defaults to a vector of 1s, i.e. no censoring.}

\item{landmark_times}{Numeric vector of length J2 giving
times at which to estimate AUC}

\item{f_hat}{Full oracle predictions (n x J1 matrix)}

\item{fs_hat}{Residual oracle predictions (n x J1 matrix)}

\item{S_hat}{Estimates of conditional event time survival function (n x J2 matrix)}

\item{G_hat}{Estimate of conditional censoring time survival function (n x J2 matrix)}

\item{folds}{Numeric vector of length n giving cross-fitting folds}

\item{sample_split}{Logical indicating whether or not to sample split}

\item{ss_folds}{Numeric vector of length n giving sample-splitting folds}

\item{time_grid_approx}{Numeric vector of length J1 giving times at which to
approximate integrals.}
}
\value{
data frame giving results
}
\description{
Estimate AUC VIM
}
