% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vim_rsquared.R
\name{vim_rsquared}
\alias{vim_rsquared}
\title{Estimate Brier score VIM}
\usage{
vim_rsquared(
  time,
  event,
  approx_times,
  landmark_times,
  f_hat,
  fs_hat,
  S_hat,
  G_hat,
  cf_folds,
  ss_folds,
  sample_split,
  scale_est = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{time}{\code{n x 1} numeric vector of observed
follow-up times If there is censoring, these are the minimum of the
event and censoring times.}

\item{event}{\code{n x 1} numeric vector of status indicators of
whether an event was observed. Defaults to a vector of 1s, i.e. no censoring.}

\item{approx_times}{Numeric vector of length J1 giving times at which to
approximate integrals.}

\item{landmark_times}{Numeric vector of length J2 giving
times at which to estimate Brier score}

\item{f_hat}{Full oracle predictions (n x J1 matrix)}

\item{fs_hat}{Residual oracle predictions (n x J1 matrix)}

\item{S_hat}{Estimates of conditional event time survival function (n x J2 matrix)}

\item{G_hat}{Estimate of conditional censoring time survival function (n x J2 matrix)}

\item{cf_folds}{Numeric vector of length n giving cross-fitting folds}

\item{ss_folds}{Numeric vector of length n giving sample-splitting folds}

\item{sample_split}{Logical indicating whether or not to sample split}

\item{scale_est}{Logical, whether or not to force the VIM estimate to be nonnegative}

\item{alpha}{The level at which to compute confidence intervals and hypothesis tests. Defaults to 0.05}
}
\value{
data frame giving results
}
\description{
Estimate Brier score VIM
}
