% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossfit_oracle_preds.R
\name{crossfit_oracle_preds}
\alias{crossfit_oracle_preds}
\title{Generate K-fold cross-fit survival predictions for downstream use}
\usage{
crossfit_oracle_preds(time, event, X, folds, pred_generator, ...)
}
\value{
data frame giving results
}
\description{
Generate K-fold cross-fit survival predictions for downstream use
}
