% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossfit_surv_preds.R
\name{crossfit_surv_preds}
\alias{crossfit_surv_preds}
\title{Generate K-fold cross-fit survival predictions for downstream use}
\usage{
crossfit_surv_preds(time, event, X, newtimes, folds, pred_generator, ...)
}
\value{
data frame giving results
}
\description{
Generate K-fold cross-fit survival predictions for downstream use
}
