% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DR_pseudo_outcome_regression.R
\name{DR_pseudo_outcome_regression}
\alias{DR_pseudo_outcome_regression}
\title{Generate K-fold cross-fit survival predictions for downstream use}
\usage{
DR_pseudo_outcome_regression(
  time,
  event,
  X,
  newX,
  S_hat,
  G_hat,
  newtimes,
  approx_times,
  SL.library
)
}
\value{
data frame giving results
}
\description{
Generate K-fold cross-fit survival predictions for downstream use
}
