% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_folds.R
\name{generate_folds}
\alias{generate_folds}
\title{Generate cross-fitting and sample-splitting folds}
\usage{
generate_folds(n, V, sample_split)
}
\value{
data frame giving results
}
\description{
Generate cross-fitting and sample-splitting folds
}
