% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_folds.R
\name{generate_folds}
\alias{generate_folds}
\title{Generate cross-fitting and sample-splitting folds}
\usage{
generate_folds(n, V, sample_split, cf_folds = NULL)
}
\arguments{
\item{n}{Total sample size}

\item{V}{Number of cross-fitting folds to use}

\item{sample_split}{Logical, whether or not sample-splitting is being used}
}
\value{
Named list of cross-fitting and sample-splitting folds
}
\description{
Generate cross-fitting and sample-splitting folds
}
