% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_delta.R
\name{p_delta}
\alias{p_delta}
\title{Wrapper for various p_delta algorithms}
\usage{
p_delta(event, X, algorithm, V = NULL, SL.library = NULL, tuning_params = NULL)
}
\arguments{
\item{event}{Indicator of event (vs censoring)}

\item{X}{Covariate matrix}

\item{algorithm}{Which binary classification algorithm to use}

\item{V}{CV fold number, required for tuned algorithms (xgboost, ranger)}

\item{SL.library}{SuperLearner library}

\item{tuning_params}{Tuning parameters for binary classification}
}
\value{
An fitted binary regression for (complement of) probability of censoring
}
\description{
Wrapper for various p_delta algorithms
}
