% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackSurv.R
\name{stackSurv}
\alias{stackSurv}
\title{Estimate a conditional survival function via stacking}
\usage{
stackSurv(
  time,
  event,
  X,
  newX,
  newtimes,
  bin_size = NULL,
  test_time = NULL,
  test_event = NULL,
  test_X = NULL,
  SL.library
)
}
\arguments{
\item{time}{Observed time}

\item{event}{Indicator of event (vs censoring)}

\item{X}{Covariate matrix}

\item{newX}{Values of covariates at which to make a prediction}

\item{newtimes}{Times at which to make the survival function prediction}

\item{bin_size}{Quantiles on which to grid times. If NULL, defaults to every observed time}

\item{SL.library}{SuperLearner library}
}
\value{
An object of class \code{stack}
}
\description{
Estimate a conditional survival function via stacking
}
