% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survML-package.R
\docType{package}
\name{survML-package}
\alias{survML}
\alias{survML-package}
\title{survML: Flexible Estimation of Conditional Survival Functions Using Regression on Observables}
\description{
An open-source R package for flexible estimation of conditional survival functions using regression on observables.
}
\author{
\strong{Maintainer}: Charles Wolock \email{cwolock@uw.edu} (\href{https://orcid.org/0000-0003-3527-1102}{ORCID})

}
\keyword{internal}
