% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survML-package.R
\docType{package}
\name{survML-package}
\alias{survML}
\alias{survML-package}
\title{survML: Tools for Flexible Survival Analysis Using Machine Learning}
\description{
Statistical methods for analyzing time-to-event data using machine learning. Implements survival stacking for conditional survival estimation and isotonic regression methods for current status data. See Wolock CJ, Gilbert PB, Simon N, and Carone M (2024) \doi{10.1080/10618600.2024.2304070}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cwolock/survML}
  \item \url{https://cwolock.github.io/survML/}
  \item Report bugs at \url{https://github.com/cwolock/survML/issues}
}

}
\author{
\strong{Maintainer}: Charles Wolock \email{cwolock@gmail.com} (\href{https://orcid.org/0000-0003-3527-1102}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Avi Kenny \email{avi.kenny@gmail.com} (\href{https://orcid.org/0000-0002-9465-7307}{ORCID}) [contributor]
}

}
\keyword{internal}
