% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survMLc.R
\name{survMLc}
\alias{survMLc}
\title{Estimate a conditional survival function using cumulative probability estimator}
\usage{
survMLc(
  time,
  event,
  X,
  newX,
  newtimes,
  time_grid_approx = sort(unique(time)),
  bin_size = NULL,
  denom_method = "stratified",
  algorithm = "xgboost",
  V = 10,
  entry = NULL,
  time_basis,
  surv_form = "PI",
  tuning_params = NULL,
  direction = "prospective",
  SL.library = NULL,
  tau = NULL
)
}
\arguments{
\item{time}{Observed time}

\item{event}{Indicator of event (vs censoring)}

\item{X}{Covariate matrix}

\item{newX}{Values of covariates at which to make a prediction}

\item{newtimes}{Times at which to make the survival function prediction}

\item{time_grid_approx}{Grid of time points on which to approximate PL integral or cumulative hazard. Defaults to observed times}

\item{bin_size}{Size of time bin on which to discretize for estimation}

\item{denom_method}{Stratified or marginal denominator}

\item{algorithm}{Which binary classification algorithm to use}

\item{V}{CV fold number, required for tuned algorithms (xgboost, ranger)}

\item{entry}{Variable indicating time of entry into the study (truncation variable) if applicable}

\item{time_basis}{How to treat time (continuous or dummy)}

\item{surv_form}{Product integral or exponential mapping}

\item{tuning_params}{Tuning parameters for binary classification}

\item{direction}{Prospective or retrospective setting}
}
\value{
An object of class \code{survMLc}
}
\description{
Estimate a conditional survival function using cumulative probability estimator
}
